Create amplification_matrix.rs
This struct represents an amplification matrix that can be applied to a 2D array of data using a convolutional network. The apply method applies the matrix to the data, while the invert method returns the inverse of the matrix if it exists. Note that the inverse is calculated using the try_inverse method of the Inverse trait, which returns None if the matrix is singular (i.e. if it cannot be inverted).